Apps/Stk : Fix for displaying idle mode text for gcf test case 27.22.4.22.1 seq1.4
When the SETUP IDLE MODE proactive command is sent the screen status request
is sent to ActivityManagerService.java. When an activity is finished or moved
to back, screen idle intent is received from the ActivityManagerService.java.
When an activity is started screen busy intent is received. If the screen is
busy the notification "idle mode text" is cancelled but is displayed when
the screen is idle.

Change-Id: Ie049bcea1001af38f53f60dd434fd86ec86a5db6
CRs-Fixed: 183837