Stk: Proper handling of high priority display text which user only can clear.
When a high priority Display Text proactive command with infinite time out is
sent,user has to clear this message. The Terminal Response for this command
will be sent from Stk framework itself. Stk app receives this message and
displays it until user clears it. In between other proactive commands like Play
Tone can be processed parallely. The Terminal Response for Play Tone command
is not sent if the Display Text message is not cleared because of improper
handling. This is corrected.

CRs-Fixed: 224350