apps/Stk: Second Display text proactive command is not displayed
Currently the STK App is sending "screen busy" terminal response
for subsequent display text proactive commands if a previous display
text is being displayed. Changes are made to update the dialog box
with the subsequent display text and not to send the screen busy.
(Refer to ETSI TS 102.384 section 27.22.4.1.4.4.2)

Change-Id: I254784a6e81c123641b836f200610a6562649da5
CRs-Fixed: 234901