These changes are related to DISPLAY Text proactive command.
The following issues related to DISPLAY_TEXT are fixed in this change:
1. Send Terminal Response if ResponseNeeded qualifier is set.
Desc: Send the Terminal Response if ResponseNeeded qualifier is set, after the
default duration else display message for ever.

2. Second Display text proactive command is not displayed.
Desc: Currently the STK App is sending "screen busy" terminal response
for subsequent display text proactive commands if a previous display
text is being displayed. Changes are made to update the dialog box
with the subsequent display text and not to send the screen busy.
(Refer to ETSI TS 102.384 section 27.22.4.1.4.4.2)

3. Proper handling of high priority display text which user only can clear
Desc: When a high priority Display Text proactive command with infinite time out is
sent,user has to clear this message. The Terminal Response for this command
will be sent from Stk framework itself. Stk app receives this message and
displays it until user clears it. In between other proactive commands like Play
Tone can be processed parallely. The Terminal Response for Play Tone command
is not sent if the Display Text message is not cleared because of improper
handling. This is corrected.

4. Do not clear Display Text with high priority, wait for user clear and no timeout.
Desc: When a high priority, wait for user clear and no timeout Display Text
Proactive command is received UE should display the text until user clears
it. While the text is being displayed if a next Proactive command is received,
it should be acted upon immediately.

Change-Id: I27e048d958d0dce27c0b513aa0ee67afd2e30f9f