Some more fixes in STK APP
1. Support to handle modem restart and ICC Refresh
Install the STK app, stop StkAppService and remove the idle text
when StkAppService receives the CARD_ABSENT/CARD_ERROR status.

Also, IdleModeText and SETUP_MENU need to be cleared when ICC status
changes to REFRESH_RESET or RESET_INIT. Hence sending the Icc Status c
hange to STkApp.

2. After Launch Browser proactive command and confirmation, the browser is
expected to go to xxx.yyy.zzz. Instead, the browser goes to the preset
home page. When sending an intent to Browser, 'action' tag also needs to
be set.

3. Default message should be displayed if SETUP_CALL message is null

4. Handle the PlayTone CMD with Null Alpha identifier

All the PlayTone CMDs are handled by the Activity intent by default.
Changing the behavior to Play the tone without creating Activity when
the PlayTone CMD is received with null Alpha identifier.

5. Include GET_CHANNEL_STATUS support in JB

Change-Id: I452781edb4dae655044536c57b69c122fee27ab5
CRs-Fixed:270755, 192553, 280258, 279494, 280254, 215985