Stk(DSDS): StkCmdReceiver should send card status to StkAppService
In case of DSDS there is only one StkAppService and when CARD_ABSENT
is received StkAppService will clean the stk icon and calls stopSelf.
System will call onDestroy during cleanup of StkAppService object.

During airplane mode whenever we received CARD_STATUS as ABSENT that will be
broadcasted to STK app service. If second card status is delayed and system
is cleaned up the StkAppService in that case StkCmdReceiver is not sending the
CARD_STATUS to StkAppService.

StkCmdReceiver need to send CARD_STATUS to StkAppService always.

Change-Id: Ie4fc0f0ba57a7f5d7d04034642062b785e71677b
CRs-fixed: 448572