Stk(DSDS): Handle the case where only one sim is inserted in DSDS
Stk icon is not shown in case of DSDS if powered up with one SIM
due to StkAppService expects two CatServices to be created.
Framework creates CatService only if card is having
atleast one application.

Remove dependency of CatService instances in StkAppService to handle
the single SIM in DSDS.

Change-Id: I868e58338434295347a5d9dde913484250163da2
CRs-fixed: 449727