Stk: Only a narrow frame will pop up in SIM toolkit main view
In StkMenuActivity.java, every time when tap menu key will
call onPrepareOptionsMenu() for showing the menu. When end
the session and back to main list, it doesn't need to show
menu after we press the menu key.

mainVisible and helpVisible variables in onPrepareOptionsMenu()
are used to judge whether to show the menu or not. So, if
the value of mainVisible || helpVisiblewe is false(ie. needn't
show menu), we should return false to throw out this event
instead of handle it.

CRs-Fixed: 533428

Change-Id: I5ac7b2f8715a08feada5423f1a52533ef78ce90d