Stk: Send StkAppService ready notification to CatService
The Android framework seems to delay the intents sent to StkApp
till the boot finishes, that is, till Boot_Complete intent
is received. As a result the proactive commands sent
before boot complete will time out.

Notify CatService that the StkAppService is ready, that is, it has
received boot_complete intent from the framework.

Change-Id: Ie52d23e698c9684c4afdf678869e41d062efcecd
Crs-Fixed: 602191