Do not inform user when the alpha identifier is a null data.
Current code is not in accordance with the spec
3GPP TS 11.14, 6.4.5. Play Tone, as user is informed when the
alpha identifier is a null data.

With reference to the spec, if there is no alpha identifier,
show the text message to the user depending on the config setting.

If alpha identifier data is null, play only tone and do not
inform the user.

Move the code to handle play tone from ToneDialog to StkAppService
in cases where only tone needs to be played and no user inform needed.

Dismiss the tone dialog from StkAppService after play tone finish.

Send the intent to StkAppService from ToneDialog to finish
the play tone, when user presses back key.

Ref spec: 3GPP TS 11.14, 6.4.5. Play Tone.

CRs-Fixed: 631723, 517578
Change-Id: I0acf11e86714138e6d3193b17181c6710d06dfb2

Send the correct Message object for the Play Tone command.

When stopping the tone that is playing due to the Play Tone command,
a delayed Message is sent to StkAppService with its 'arg2' field
containing a flag that indicates if tone dialog activity needs to be
closed. But when handling the received messages, StkAppService expects
the Message object's 'arg2' field to contain the slot Id on which the
command is received. This causes unexpected behavior when executing
the Play Tone command.

Use a different field in the Message object to send the the value of
the flag which indicates if tone dialog activity needs to be closed.

Change-Id: Ib25986efe1d7c26976015f666580a3deb28bb328
CRs-Fixed: 915779

Send slotId in message for OP_STOP_TONE

- unable to send the response to the
correct sim card
- Set the slotId as arg2 in the message

Change-Id: I31143b35af5d19077eb7224736aef9c664016246
CRs-Fixed: 989267

Change-Id: Ib25986efe1d7c26976015f666580a3deb28bb328