Handle the timer in StkMenuActivity correctly.
Currently the timer in StkMenuActivity is started only
if the state is SECONDARY.

While handling multiple back to back GET INPUT screens, the
timer in StkMenuActivity gets started before setting the state
as MAIN, resulting in the GET INPUT screen switch to main menu.
Start the timer in StkMenuActivity after setting the state to MAIN
for cases like back to back input screens.

Change-Id: Ieee89429b44ead0173d7039703b1b495183e54cb
CRs-Fixed: 651303