STK: Close the stk dialog when StkDialogActivity's onStop is called
When the StkDialogActivity is put in the background, its onStop
callback method is called by the framework. The StkDialogActivity
sends RES_ID_TIMEOUT response, but the dialog still remains in the
background.

Close the stk dialog and cancel the timer after the RES_ID_TIMEOUT
is sent when the acitvity is put in the background.

Change-Id: Ifc19cfc0765e094ec1f742754926edae949e860e
CRs-Fixed: 879685