android:watch Fix of phone app crash because of null point exeption in stkDialogActivity
As because of NoActionbar theme in watch, in StkDialogActivity title bar with icon will not be initialized.
As a result in stkDialogActivity, we are trying set  image for uninitialized image of title bar, which result crash.
In order to fix this, getActionBar() == null check added before setting image of titlebar.

Change-Id: I10c32ec1d778e430a55ec7e19913d89810600d19